Pass cloner factory to bind from outside
[#85215950]
[#86816248]
